ansible/roles/elasticsearch/files/metax-refdata-indexer/requirements.txt
requests
elasticsearch
unicode-slugify
